fix tests now that pseudo is around
